Modal folder reorganized
