+++ src/components/Modal/FollowUpModal/FollowUpModal.tsx
@@ -1,18 +1,19 @@
-import React, { useCallback, useMemo } from 'react'
-import { Modal, ModalProps } from 'decentraland-ui/dist/components/Modal/Modal'
-import { Header } from 'decentraland-ui/dist/components/Header/Header'
-import { Close } from 'decentraland-ui/dist/components/Close/Close'
-import { Button } from 'decentraland-ui/dist/components/Button/Button'
-import TokenList from 'decentraland-gatsby/dist/utils/dom/TokenList'
-import useClipboardCopy from 'decentraland-gatsby/dist/hooks/useClipboardCopy'
-
-import './ProposalModal.css'
 import './FollowUpModal.css'
+import '../ProposalModal.css'
+
 import Paragraph from 'decentraland-gatsby/dist/components/Text/Paragraph'
+import useClipboardCopy from 'decentraland-gatsby/dist/hooks/useClipboardCopy'
 import useFormatMessage from 'decentraland-gatsby/dist/hooks/useFormatMessage'
 import Time from 'decentraland-gatsby/dist/utils/date/Time'
-import { ProposalAttributes } from '../../entities/Proposal/types'
-import { proposalUrl, forumUrl, JOIN_DISCORD_URL } from '../../entities/Proposal/utils'
+import TokenList from 'decentraland-gatsby/dist/utils/dom/TokenList'
+import { Button } from 'decentraland-ui/dist/components/Button/Button'
+import { Close } from 'decentraland-ui/dist/components/Close/Close'
+import { Header } from 'decentraland-ui/dist/components/Header/Header'
+import { Modal, ModalProps } from 'decentraland-ui/dist/components/Modal/Modal'
+import React, { useCallback, useMemo } from 'react'
+
+import { ProposalAttributes } from '../../../entities/Proposal/types'
+import { forumUrl, JOIN_DISCORD_URL, proposalUrl } from '../../../entities/Proposal/utils'
 
 export type FollowUpModalProps = Omit<ModalProps, 'children'> & {
   onDismiss: (e: React.MouseEvent<any>) => void
