+++ src/pages/proposal.tsx
@@ -1,49 +1,49 @@
-import React, { useMemo, useEffect } from 'react'
-import { useLocation } from '@gatsbyjs/reach-router'
-import { Personal } from 'web3x/personal'
-import { Address } from 'web3x/address'
-import useAsyncTask from 'decentraland-gatsby/dist/hooks/useAsyncTask'
-import useAsyncMemo from 'decentraland-gatsby/dist/hooks/useAsyncMemo'
-import usePatchState from 'decentraland-gatsby/dist/hooks/usePatchState'
-import Grid from 'semantic-ui-react/dist/commonjs/collections/Grid/Grid'
-import { Header } from 'decentraland-ui/dist/components/Header/Header'
-import { Loader } from 'decentraland-ui/dist/components/Loader/Loader'
-import { Button } from 'decentraland-ui/dist/components/Button/Button'
-import { navigate } from 'decentraland-gatsby/dist/plugins/intl'
-import { Governance } from '../api/Governance'
-import useProposal from '../hooks/useProposal'
-
-import ContentLayout, { ContentSection } from '../components/Layout/ContentLayout'
-import CategoryLabel from '../components/Category/CategoryLabel'
-import StatusLabel from '../components/Status/StatusLabel'
-import ForumButton from '../components/Section/ForumButton'
-import SubscribeButton from '../components/Section/SubscribeButton'
-import ProposalResultSection from '../components/Section/ProposalResultSection'
-import ProposalDetailSection from '../components/Section/ProposalDetailSection'
-import useAuthContext from 'decentraland-gatsby/dist/context/Auth/useAuthContext'
-import { forumUrl } from '../entities/Proposal/utils'
-import { Snapshot } from '../api/Snapshot'
-import Markdown from 'decentraland-gatsby/dist/components/Text/Markdown'
-import { VoteRegisteredModal } from '../components/Modal/VoteRegisteredModal'
-import { DeleteProposalModal } from '../components/Modal/DeleteProposalModal'
-import { VotesList } from '../components/Modal/VotesList'
-import useFormatMessage from 'decentraland-gatsby/dist/hooks/useFormatMessage'
-import retry from 'decentraland-gatsby/dist/utils/promise/retry'
-import locations from '../modules/locations'
-import { UpdateProposalStatusModal } from '../components/Modal/UpdateProposalStatusModal'
-import { ProposalStatus } from '../entities/Proposal/types'
-import ProposalHeaderPoi from '../components/Proposal/ProposalHeaderPoi'
-import Head from 'decentraland-gatsby/dist/components/Head/Head'
-import { formatDescription } from 'decentraland-gatsby/dist/components/Head/utils'
-
 import './index.css'
 import './proposal.css'
-import NotFound from 'decentraland-gatsby/dist/components/Layout/NotFound'
-import ProposalFooterPoi from '../components/Proposal/ProposalFooterPoi'
-import ProposalComments from '../components/Proposal/ProposalComments'
-import { FollowUpModal } from '../components/Modal/FollowUpModal'
-import { isUnderMaintenance } from '../modules/maintenance'
+
+import { useLocation } from '@gatsbyjs/reach-router'
+import Head from 'decentraland-gatsby/dist/components/Head/Head'
+import { formatDescription } from 'decentraland-gatsby/dist/components/Head/utils'
 import MaintenancePage from 'decentraland-gatsby/dist/components/Layout/MaintenancePage'
+import NotFound from 'decentraland-gatsby/dist/components/Layout/NotFound'
+import Markdown from 'decentraland-gatsby/dist/components/Text/Markdown'
+import useAuthContext from 'decentraland-gatsby/dist/context/Auth/useAuthContext'
+import useAsyncMemo from 'decentraland-gatsby/dist/hooks/useAsyncMemo'
+import useAsyncTask from 'decentraland-gatsby/dist/hooks/useAsyncTask'
+import useFormatMessage from 'decentraland-gatsby/dist/hooks/useFormatMessage'
+import usePatchState from 'decentraland-gatsby/dist/hooks/usePatchState'
+import { navigate } from 'decentraland-gatsby/dist/plugins/intl'
+import retry from 'decentraland-gatsby/dist/utils/promise/retry'
+import { Button } from 'decentraland-ui/dist/components/Button/Button'
+import { Header } from 'decentraland-ui/dist/components/Header/Header'
+import { Loader } from 'decentraland-ui/dist/components/Loader/Loader'
+import React, { useEffect, useMemo } from 'react'
+import Grid from 'semantic-ui-react/dist/commonjs/collections/Grid/Grid'
+import { Address } from 'web3x/address'
+import { Personal } from 'web3x/personal'
+
+import { Governance } from '../api/Governance'
+import { Snapshot } from '../api/Snapshot'
+import CategoryLabel from '../components/Category/CategoryLabel'
+import ContentLayout, { ContentSection } from '../components/Layout/ContentLayout'
+import { DeleteProposalModal } from '../components/Modal/DeleteProposalModal/DeleteProposalModal'
+import { FollowUpModal } from '../components/Modal/FollowUpModal/FollowUpModal'
+import { UpdateProposalStatusModal } from '../components/Modal/UpdateProposalStatusModal/UpdateProposalStatusModal'
+import { VoteRegisteredModal } from '../components/Modal/Votes/VoteRegisteredModal'
+import { VotesList } from '../components/Modal/Votes/VotesList'
+import ProposalComments from '../components/Proposal/ProposalComments'
+import ProposalFooterPoi from '../components/Proposal/ProposalFooterPoi'
+import ProposalHeaderPoi from '../components/Proposal/ProposalHeaderPoi'
+import ForumButton from '../components/Section/ForumButton'
+import ProposalDetailSection from '../components/Section/ProposalDetailSection'
+import ProposalResultSection from '../components/Section/ProposalResultSection'
+import SubscribeButton from '../components/Section/SubscribeButton'
+import StatusLabel from '../components/Status/StatusLabel'
+import { ProposalStatus } from '../entities/Proposal/types'
+import { forumUrl } from '../entities/Proposal/utils'
+import useProposal from '../hooks/useProposal'
+import locations from '../modules/locations'
+import { isUnderMaintenance } from '../modules/maintenance'
 
 type ProposalPageOptions = {
   changing: boolean
