+++ src/pages/submit/index.tsx
@@ -1,15 +1,16 @@
-import React, { useState } from 'react'
-import { Header } from 'decentraland-ui/dist/components/Header/Header'
+import './submit.css'
 
-import useFormatMessage from 'decentraland-gatsby/dist/hooks/useFormatMessage'
-import ContentLayout, { ContentSection } from '../../components/Layout/ContentLayout'
 import Head from 'decentraland-gatsby/dist/components/Head/Head'
 import Paragraph from 'decentraland-gatsby/dist/components/Text/Paragraph'
+import useFormatMessage from 'decentraland-gatsby/dist/hooks/useFormatMessage'
+import { Header } from 'decentraland-ui/dist/components/Header/Header'
+import React, { useState } from 'react'
+
 import CategoryBanner from '../../components/Category/CategoryBanner'
+import ContentLayout, { ContentSection } from '../../components/Layout/ContentLayout'
+import { POIProposalModal } from '../../components/Modal/POIProposalModal/POIProposalModal'
 import { ProposalType } from '../../entities/Proposal/types'
 import locations from '../../modules/locations'
-import './submit.css'
-import { POIProposalModal } from '../../components/Modal/POIProposalModal'
 
 export default function NewProposalPage() {
   const t = useFormatMessage()
