+++ tsconfig.json
@@ -17,5 +17,6 @@
     "typeRoots": ["node_modules/@types", "src/@types"] /* List of folders to include type definitions from. */,
     "esModuleInterop": true /* Enables emit interoperability between CommonJS and ES Modules via creation of namespace objects for all imports. Implies 'allowSyntheticDefaultImports'. */
   },
-  "include": ["src"]
+  "include": ["src"],
+  "exclude": ["node_modules", "**/node_modules/*"]
 }
